Adding db config and product add for admin

+++ StoreManagement/src/main/java/org/example/domain/Product.java
@@ -8,13 +8,13 @@ public class Product {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     @NotBlank private String name;
-    @Min(0) private double price;
+    @Min(0) private int price;
     public Product() {}
-    public Product(String name, double price) {this.name=name;this.price=price;}
+    public Product(String name, int price) {this.name=name;this.price=price;}
     public Long getId() {return id;}
     public void setId(Long id) {this.id=id;}
     public String getName() {return name;}
     public void setName(String name) {this.name=name;}
-    public double getPrice() {return price;}
-    public void setPrice(double price) {this.price=price;}
+    public int getPrice() {return price;}
+    public void setPrice(int price) {this.price=price;}
 }
